Simplifying and fixing NbTopManager & Lkp initialization and synchronization. In TopManager, getDefault used double-checked locking. This was carried over into NbTopManager after #19443, which mostly worked. However XTest was broken since NbTM.get() did not guarantee that the returned instance would actually have finished initializing (loading modules, etc.), and there was no alternative way of guaranteeing that. Consequently XTest code would see a half-started NB installation and behave erratically (in ide mode - code mode was fine).
Now get() blocks until the NbTM is done. If you don't want that, and
you understand that you might be getting a semiconstructed instance,
use getUninitialized(). This is useful currently in a few places where
some code calls back into NbTM during startup, and there is some
protected method that the NbTM impl can select among.

In the future this getUninitialized() code pattern should be removed.
For each discrete block of configurable functionality (e.g.: window
sys init mode; error stream selection; etc.) there should be a manager
class accessible in Lookup to implement it. Then the different NbTM
subclasses should be deleted (see also #20253). Style of running
should be configured by including appropriate JARs (with
META-INF/services/*) in the classpath: e.g. the GUI support JAR, or
the scripting support JAR.

On the Lkp side, there was a bunch of weird code involving spawning
threads to init lookup and then joining with them. Deleted all that.
When moduleClassPathInitialized is called, Services/ is added to the
lookup pool together with other stuff. #25917 is not likely to recur,
at least not in that form; other things have changed, incl. correction
of MetaInfServicesLookup synchronization.