Fix for issue 35178, no accessible description for tree in diff engine, etc.
Also fixed a rather evil listener leak (in cooperation with a similar
leak in PropertySheet, invoking the editor twice resulted in a
value permanently slowly toggling between two values forever).

Provided a workaround for the Diff module to not throw an NPE when
a selection is made (it writes to files in the folder the editor is
showing, causing the tree to get rebuilt;  BeanTreeView posts a
runnable to the event thread to clear the selection briefly, which
ObjectEditor interprets as the user setting the value to null,
which the diff settings class can't handle).

Note a strong warning should be made that classes using this editor
should NOT modify files in the folder the editor tree is viewing, when
the selection changes.  It also causes a problem that the ordering
of nodes in the editor changes when it is clicked.

All in all, the maintenance cost of this class far exceeds its
(very) marginal utility.

I suggest replacing the Explorer-based UI with a set of radio buttons.
We could do a better UI (include the node description of each option),
and would not have the (severe) event thrashing this editor causes.
Or kill the thing entirely.