Upgrading the Jasper libraries used by the JSP parser to Tomcat 5.0.16.
The change was non-trivial, as there are some subtle changes between the previous version (5.0.9) and this one. Had to do some fixes and improvements in the classloading architecture of Jasper.

Also updating the test application (jsp-exampes) to use the version from Tomcat 5.0.16.
The test golden files were changed slightly, but the changes are reasonable.

This also upgrades the servlet and JSPs APIs to an official version approved by the JSRs.