Redirecting the output of the Tomcat process to the IDE output window. This is done as follows:
There is a general reusable class called ProcessSupport that redirects outputs of any process to the IDE output window. This class could eventually be contributed to OpenIDE.
Next, the StartTomcat class uses ProcessSupport to redirect the Tomcat process' output to the IDE output window. A change needed to be done in the way Tomcat starts: by calling "tomcat.sh run" rather than "startup.sh" (same applies for the corresponding .bat counterparts on Windows), Tomcat does no longer start a separate window.

Also an unrelated fix in that the monitor configuration synchronization should only be done when starting Tomcat, not when stopping it.