Fix a long known issue with VW (and Creator) about requiring closing and reopening of projects after making changes to and compiling non life cycle bean classes in the project.
The implementation works like this:

I have implemented a ProjectBuiltQuery which is anlogue of FileBuiltQuery at the project level.
The basic idea is that the project is considered built if all the files are built (i.e. compiled).
When the project's built state changes an event is fired. Insync reacts to the transition of the
project state from non-built to built state and recreates the internal Project Class Loader.
Subsequently the user needs to only hit Refresh button in the designer toolbar to pick up the
changes made to the compiled file. No need to close project.

I will look into getting rid of the need to hit the Refresh button also.