Bug Fix# 104668 components disappear when editing label
Do not use the skipCount field to determine how many elements
of normal rendered output of the in-place edited component, as
it was initialized to the count of children of the prerendered fragment.

NOTE: The prerendered fragment is used for in-place editing
purposes in place of normal rendered output of the component.