Bug Fix # 114783]  Exceptions thrown when modifying re-opened project
Problem: The problem was due to the fact that a DocumentListener was being registered twice. When Insync made modifications - the second of them was removed (using the "this" reference) while the modifications were being made. However the first  listener still was present which tried to mark the source dirty, hence the exception. The double registration occurred due to a race condition between property change listener fired to indicate document open and document access from editor cookie and registration of the listener on a non null document receipt. Ironically the condition occured more often on a single processor systems. This issue came to fore because of  a recent fix to JsfJspDataObject EditorCookie #105318 which in tandem with issue that always opens the JSP source tab of the designer (issue #?).

Fix: The fix is to prevent duplicate registration of the listeners.