#130869: Reverting back the old behaviour to open everything in AWT
However guys are you sure that your code is correct? You are inspecting the AWT hierarchy trying to find some component and I really do not like it. It reminds me reflection, and it is not good habbit to use reflection. Instead you should modify the API in CloneableEditor to give you what you need without these hacks.