Initial implementation of Hibernate Reverse Engineering feature.
This commit includes a database tables panel, to show the database tables based on the db selection.
DB selection is temporary till I get a service api to get the configuration files with proper db information.

It also contains a code generation panel, which is used to pick a package name for the generated pojos and hbm files.
Generation of Pojos and hbm files are not done in this commit.

It also contains some of the common classes from J2ee persistence to show the database tables based on the database selection.
I copied them from j2ee.persistence and modified them to work with Hibernate.
That way Hibernate doesn't have to completely depend on j2ee.persistence.

With this commit, one can create a simple hibernate.reveng.xml file by selecting some tables
from the database tables panel.