Fixed Issue 138010 - package has a very tall tab after entering a stereotype and tagged value
We are not behaving like 6.1.  Only the name is added to the tab.  The stereotype and tagged values will remain in the body.  We are behaving better than 6.1, because the tagged values use to always appear at the bottom of the package, now the stereotype, name and tagged values will always stay together.