Due to recent update to Firefox 2.0.0.16...the debugger extension was launching two browser windows -
netbeans javascript extension splah screen window
debugee application window

this was caused because of the way Firefox is invoked (first time) and the way the URL is passed in
the arguments property of the XUL window object. Had to change the strategy to determine when to
lauch a new window and when to reuse the current window.