Bug Fix  142593 Is not possible to install bundled extension for Firefox 2.0
When the profile is originally created by Firefox 3 and then reused by Firefox 2 (which
will be the case on Ubuntu 8.04 as the user has to downgrade the Firefox version to be
able to use the NetBeans JavaScript debugger) the Firefox 2 is not noticing the
newly installed Firebug and NetBeans JavaSCript debugger extension. Removing the file
named extensions.rdf does the trick.

If the user does not install the package firefox-2-gnome-support, in that case
access to nsIShellService service of Firefox throws an exception and needs to be
protected with try/catch block.