Issue 139496 - relationship discovery does not restore containment link and role binding
Nested Link - The problem was that when you called delete on a proxy presentation element it was deleting the element it was proxying.  That was not correct, only the proxy should be deleted.
Role Binding - I also was not able to reproduce this problem.
Comment Links - When you deleted a comment link we where always removing the relationship from the model.