Repackage the modified Ruby and JavaScript parsers in the IDE from their original package names to packagename+".nb". This is done to avoid unintentional interactions with the standard versions of these parsers. For example, attempts to use scripting engines (such as JRuby) has reportedly had problems (P2 #121952 for Ruby, and some reports about Mozilla Rhino as well. Another possible bug this fixes is 144637, or at least exception 91820 submitted as a duplicate of that bug.)
The repackaging is done as token replacements as the JRuby and Rhino
parser source zip files are extracted. It turns out JRuby had split
two of its packages into separate jars: the bytelist library, and
JNA-posix. These are left in the original package, and I extracted the
bytelist and jvyamlb libraries into proper NetBeans libraries (libs.bytelist
and libs.jvyamlb). This will help the upcoming task of removing the
dependency of the YAML editor on the JRuby distribution. YAML is used
with Python as well (and in particular in the Google App Engine) so
we should be able to distribute it with the Python distro without JRuby.