More support for 1.9 debugging:
Try whether there is no ruby-debug-ide19 gem when we are running on 1.9. This
is supposed to be rather temporary solution. In ideal case we would have
ruby-debug-ide working with both Ruby version, 1.8 and 1.9. Let's go with this
compromise for now and let the user choose (by having installed appropriate
gem) what should be used.