Issue #201612 - honor resolution:=optional when creating catalogue entry from osgi bundle
hotfix #09d776d9da10 caused that in the uc catalog is no dependency declared at all between jsch and jzlib. Even if it is only optional, for the need of netbeans it makes more sense to declare it as mandatory, so that jzlib will be always downloaded together with jsch.