#219451 - OutOfMemoryError: Java heap space Two copies of parsed source code are not held by Context anymore (source and sanitizedSource).
Sanitized source is replaced by SanitizedPart, which holds just a "diff to original source" after sanitization.

Common original source is abstracted by SourceHolder with two implementations - SnapshotSourceHolder holds just a snapshot and returns an original source code directly from it and StringSourceHolder holds a whole source code. That could be improved as well in the future.