Refactoring of resolving function return types.
Semi types with variables are resolved lazily in function scope so it doesn't have to be counted in Model visitor. Offset for resolving types is counted from function elements (the ending offset of the last one is used), because function offset (start of the name) can't be used to solve variable types which are defined after that function offset (e.g. return statement "return $foo;" is at the end of the function body). If function doesn't contain any element, then function offset is used.