#240156: Do not update cached names when refreshing not existing children
ChildrenSupport.notExistingChildren can contain files with incorrect (not canonical) names.
If they are detected as valid (if they really exists, but with different canonical name), the
cached name is updated to the incorrect case. We need a way to tell the NamingFactory that
the passed file name is not guaranteed to be canonical and that the cached name should not
be changed.