#224986 - Stash functionality
Stashes are now displayed in the repository browser view under the Stashes node.
Stashes may be saved either from the repo browser when selecting Stash changes from
the popup on the Stashes node or by invoking Stash Changes action from the Quick Search view.
To apply a stash it's enough to dbl-click on its relevant node and to drop the stash simple Del
is available on the node.